feat(spotify): add a provider class to provide tokens for access the Spotify Rest API
The SpotifyAuthorizationProvider class access the Spotify API to get a access token.
The generated token have not user data, because the grant type is client_credentials,
then this token cannot be used to access user data.

This token will be cached in memory while it is not expired.

See the Spotify documentation: https://developer.spotify.com/documentation/general/guides/authorization-guide/#client-credentials-flow